% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios2indicators.R
\encoding{UTF-8}
\name{esios2indicators}
\alias{esios2indicators}
\title{Download ESIOS indicators metadata and augment with units}
\usage{
esios2indicators(api_key, output_dir = ".", encoding = "UTF-8")
}
\arguments{
\item{api_key}{String: your ESIOS API key.}

\item{output_dir}{Directory where files will be saved (default ".").}

\item{encoding}{Encoding used to read the JSON (default "UTF-8").}
}
\value{
Invisible data.frame with columns \code{id}, \code{short_name_es},
  \code{name_es}, \code{short_name_en}, \code{name_en}, and \code{units}.
}
\description{
\code{esios2indicators} fetches the list of available indicators
from Spain's ESIOS API in both Spanish and English, cleans their
short names (replacing spaces or hyphens with "_"), appends the ID
to ensure uniqueness, assigns units manually, and saves
the result to CSV and RDA files.
}
