% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios2lfdata.R
\encoding{UTF-8}
\name{esios2lfdata}
\alias{esios2lfdata}
\title{Build daily curves into \code{ldata} object from ESIOS data}
\usage{
esios2lfdata(
  var_names,
  start_date,
  end_date,
  api_key,
  resolution = c("1min", "5min", "10min", "15min", "30min", "hour", "1hour", "4hour",
    "1day"),
  verbose = FALSE
)
}
\arguments{
\item{var_names}{Character vector of indicator names in "ShortName_ID" format.}

\item{start_date}{String "YYYY-MM-DD" interval start (Europe/Madrid).}

\item{end_date}{String "YYYY-MM-DD" interval end (Europe/Madrid).}

\item{api_key}{Your ESIOS API key.}

\item{resolution}{Character: one of
\code{"1min","5min","10min","15min","30min",
"hour","1hour","4hour","1day"}.}

\item{verbose}{Logical; if \code{TRUE}, prints progress messages.}
}
\value{
A \code{ldata} list: element \code{df} is a data.frame (index, date, leap, dst),
  followed by each variable as a daily \code{fdata} object.
}
\description{
Fetches ESIOS time-series for a specified resolution and date range via \code{esios2df()},
  then assembles a single \code{ldata} list: common metadata (daily index, date, leap, dst)
  plus one \code{fdata} per variable with daily curves.
}
